requirements.txt
numpy
matplotlib
keras
tensorflow
opencv-python
scikit-learn
pandas
scikit-image
seaborn
h5py
scipy
pillow
requests
jupyter